commit after finish scheduler With O(1) for Lab4

+++ STM32_cube_IDE/Core/Inc/global.h
@@ -9,12 +9,14 @@
 #define INC_GLOBAL_H_
 
 #include "main.h"
+#include <stdlib.h>
+#include <stdint.h>
 
 #define TICK 10
 
 #define true 1
 #define false 0
-typedef int bool;
+typedef uint8_t bool;
 
 void blink_led();
 
@@ -26,5 +28,6 @@ void blink_led();
 #include "fsm_automatic.h"
 #include "fsm_manual.h"
 #include "fsm_setting.h"
+#include "Scheduler.h"
 
 #endif /* INC_GLOBAL_H_ */
